feat: wallmount (#14)
## 🧰 Wallmount for HomeRacker
A new utility for your modular setup!

With this accessory, you can mount a HomeRacker frame to a wall, the
floor — or even the ceiling if you're feeling adventurous.

### 🔧 Customizable Bore Holes
You can choose between:
* Countersunk Holes
* Flathead Holes

For both options, you can specify the shaft diameter to match your
screws.
For the countersunk option, you can also set the countersink angle — the
angle at which the screw head tapers.
By default, this is set to 90°. 📐